backend/requirements.txt
fastapi
uvicorn
sqlmodel
pandas
python-dotenv
requests
python-multipart
psycopg2-binary
python-jose[cryptography]
slowapi
pydantic[email]
